Test each class in its own file
