+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 require "bundler/setup"
 require "carto/common/encryption_service"
 require "carto/common/logger_formatter"
+require 'google/cloud/pubsub'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -13,3 +14,36 @@ RSpec.configure do |config|
     c.syntax = :expect
   end
 end
+
+# See https://relishapp.com/rspec/rspec-mocks/v/3-0/docs/verifying-doubles/dynamic-classes
+# rubocop:disable Lint/UselessMethodDefinition
+class PubsubDouble
+
+  include Google::Cloud::Pubsub
+
+  def create_topic(*)
+    super
+  end
+
+  def get_topic(*)
+    super
+  end
+
+  def get_subscription(*)
+    super
+  end
+
+end
+
+class PubsubMessageDouble < Google::Cloud::Pubsub::Message
+
+  def ack!(*)
+    super
+  end
+
+  def reject!(*)
+    super
+  end
+
+end
+# rubocop:enable Lint/UselessMethodDefinition
